Bootstrap the agent-api SDK generation pipeline
Flatten the repo to root and ship the OpenAPI -> Python SDK generation
pipeline plus public-repo essentials. The SDK itself is generated, not
hand-written: the live agent_platform public v2 schema is the single source
of truth, and src/ is committed by automated sync PRs (regen-on-dispatch),
not in this bootstrap.

- Generation: prepare_openapi.py (codegen normalization only), config.yaml,
  templates/, openapi.json contract snapshot.
- Workflows: regen-on-dispatch (auto-PR on backend changes, with SemVer
  bump), nightly drift check, PR test (regenerates + tests), CodeQL,
  uv.lock private-ref guard.
- OSS essentials: LICENSE (MIT), CODEOWNERS, SECURITY, CONTRIBUTING,
  dependabot.

Version starts at 0.0.0; the first sync PR cuts a clean 0.1.0.
