+++ .github/CODEOWNERS
@@ -1,0 +1,3 @@
+# Default code owners for every file. Pull requests auto-request review from
+# both unless explicitly overridden by a more specific pattern below.
+* @abonneth @adeprezh
