+++ .github/workflows/nightly-drift-check.yaml
@@ -1,10 +1,9 @@
 name: Nightly drift check
 
-# Safety net for the dispatch pipeline. Fetches the live production OpenAPI
-# schema from https://agp.hcompany.ai/openapi.json, regenerates the SDK, and
+# Safety net for the dispatch pipeline. Fetches the most recent OpenAPI schema
+# artifact produced by agent_platform's `main` branch, regenerates the SDK, and
 # opens a PR if the committed src/ has drifted. Catches missed dispatches
-# (e.g., agent_platform's schema-sync workflow failed silently, or the schema
-# changed via an out-of-band deploy).
+# (e.g., agent_platform's schema-sync workflow failed silently to dispatch).
 
 on:
   schedule:
@@ -12,9 +11,6 @@ on:
     # agent_platform, so any breakage there surfaces first.
     - cron: "0 4 * * *"
   workflow_dispatch:
-
-env:
-  OPENAPI_URL: "https://agp.hcompany.ai/openapi.json"
 
 permissions:
   contents: write
@@ -37,24 +33,42 @@ jobs:
       - name: Install openapi-python-client
         run: uv tool install openapi-python-client==0.28.3
 
-      - name: Fetch live OpenAPI schema
+      - name: Fetch latest OpenAPI artifact from agent_platform main
+        env:
+          GH_TOKEN: ${{ secrets.AGP_ARTIFACT_TOKEN }}
         run: |
-          curl --fail --silent --show-error "$OPENAPI_URL" -o /tmp/openapi.json
+          # Find the most recent successful run of the schema-export workflow on main.
+          run_id=$(gh run list \
+            --repo hcompai/agent_platform \
+            --branch main \
+            --workflow sdk-sync-dispatch.yaml \
+            --status success \
+            --limit 1 \
+            --json databaseId \
+            --jq '.[0].databaseId')
+          if [ -z "$run_id" ] || [ "$run_id" = "null" ]; then
+            echo "::error::No successful sdk-sync-dispatch run found on agent_platform main"
+            exit 1
+          fi
+          echo "Pulling openapi.json from agent_platform run $run_id"
+          gh run download "$run_id" \
+            --repo hcompai/agent_platform \
+            --name openapi-schema \
+            --dir /tmp/agp-schema
           python3 - <<'PY'
           import json
-          with open("/tmp/openapi.json") as f:
+          with open("/tmp/agp-schema/openapi.json") as f:
               schema = json.load(f)
           assert schema.get("openapi", "").startswith("3."), f"Not OpenAPI 3.x: {schema.get('openapi')!r}"
           paths = schema.get("paths", {})
-          assert len(paths) > 5, f"Suspiciously few paths in live schema: {len(paths)}"
-          print(f"Live schema: openapi={schema['openapi']}, {len(paths)} paths")
+          assert len(paths) > 5, f"Suspiciously few paths in fetched schema: {len(paths)}"
+          print(f"Schema OK: openapi={schema['openapi']}, {len(paths)} paths")
           PY
 
-      - name: Regenerate SDK from live schema
-        working-directory: python
+      - name: Regenerate SDK from fetched schema
         run: |
-          uv run --no-project --with pyyaml python scripts/filter_openapi.py \
-            /tmp/openapi.json filter.yaml /tmp/openapi.filtered.json
+          uv run --no-project python scripts/prepare_openapi.py \
+            /tmp/agp-schema/openapi.json /tmp/openapi.filtered.json
           rm -rf src/agent_platform/
           openapi-python-client generate \
             --path /tmp/openapi.filtered.json \
@@ -67,7 +81,7 @@ jobs:
           rm /tmp/openapi.filtered.json
 
       - name: Refresh openapi.json snapshot
-        run: cp /tmp/openapi.json openapi.json
+        run: cp /tmp/agp-schema/openapi.json openapi.json
 
       - name: Detect drift
         id: diff
@@ -75,7 +89,7 @@ jobs:
           git add -A
           if git diff --staged --quiet; then
             echo "drift=false" >> "$GITHUB_OUTPUT"
-            echo "No drift detected. Live schema matches committed snapshot."
+            echo "No drift detected. Committed snapshot matches latest agent_platform main."
           else
             echo "drift=true" >> "$GITHUB_OUTPUT"
             echo "## Drift detected" >> "$GITHUB_STEP_SUMMARY"
@@ -89,17 +103,17 @@ jobs:
           token: ${{ secrets.GITHUB_TOKEN }}
           branch: drift/nightly-${{ github.run_id }}
           delete-branch: true
-          commit-message: "chore(sdk): nightly drift sync against live schema"
-          title: "chore(sdk): nightly drift sync against live schema"
+          commit-message: "chore(sdk): nightly drift sync against agent_platform main"
+          title: "chore(sdk): nightly drift sync against agent_platform main"
           labels: auto-sync,drift
           body: |
             The nightly safety-net workflow detected drift between the committed
-            `python/src/` (+ `openapi.json` snapshot) and the live schema at
-            `${{ env.OPENAPI_URL }}`.
+            `src/` (+ `openapi.json` snapshot) and the latest OpenAPI schema
+            artifact published by `hcompai/agent_platform`'s `main` branch.
 
             This usually means the `repository_dispatch` from `hcompai/agent_platform`
-            was missed (e.g., transient failure, or a backend deploy that bypassed
-            the normal merge path). The regeneration here closes the gap.
+            was missed (transient failure, or a force-push / out-of-band update).
+            The regeneration here closes the gap.
 
             If this PR is opened repeatedly, investigate the dispatch path in
-            `hcompai/agent_platform/.github/workflows/`.
+            `hcompai/agent_platform/.github/workflows/sdk-sync-dispatch.yaml`.
