+++ .github/workflows/regen-on-dispatch.yaml
@@ -59,21 +59,25 @@ jobs:
           # Unify repository_dispatch and workflow_dispatch into one set of step outputs
           # so subsequent steps don't have to branch on event type.
           if [ "${{ github.event_name }}" = "workflow_dispatch" ]; then
-            echo "agp_sha=${{ inputs.agp_sha }}" >> "$GITHUB_OUTPUT"
-            echo "agp_short_sha=$(echo "${{ inputs.agp_sha }}" | cut -c1-7)" >> "$GITHUB_OUTPUT"
-            echo "agp_actor=manual:${{ github.actor }}" >> "$GITHUB_OUTPUT"
-            echo "agp_run_id=${{ inputs.agp_run_id }}" >> "$GITHUB_OUTPUT"
-            echo "agp_artifact_name=${{ inputs.agp_artifact_name }}" >> "$GITHUB_OUTPUT"
-            echo "agp_commit_message=manual regen" >> "$GITHUB_OUTPUT"
-            echo "breaking=false" >> "$GITHUB_OUTPUT"
+            {
+              echo "agp_sha=${{ inputs.agp_sha }}"
+              echo "agp_short_sha=$(echo "${{ inputs.agp_sha }}" | cut -c1-7)"
+              echo "agp_actor=manual:${{ github.actor }}"
+              echo "agp_run_id=${{ inputs.agp_run_id }}"
+              echo "agp_artifact_name=${{ inputs.agp_artifact_name }}"
+              echo "agp_commit_message=manual regen"
+              echo "breaking=false"
+            } >> "$GITHUB_OUTPUT"
           else
-            echo "agp_sha=${{ github.event.client_payload.agp_sha }}" >> "$GITHUB_OUTPUT"
-            echo "agp_short_sha=${{ github.event.client_payload.agp_short_sha }}" >> "$GITHUB_OUTPUT"
-            echo "agp_actor=${{ github.event.client_payload.agp_actor }}" >> "$GITHUB_OUTPUT"
-            echo "agp_run_id=${{ github.event.client_payload.agp_run_id }}" >> "$GITHUB_OUTPUT"
-            echo "agp_artifact_name=${{ github.event.client_payload.agp_artifact_name }}" >> "$GITHUB_OUTPUT"
-            echo "agp_commit_message=${{ github.event.client_payload.agp_commit_message }}" >> "$GITHUB_OUTPUT"
-            echo "breaking=${{ github.event.client_payload.breaking }}" >> "$GITHUB_OUTPUT"
+            {
+              echo "agp_sha=${{ github.event.client_payload.agp_sha }}"
+              echo "agp_short_sha=${{ github.event.client_payload.agp_short_sha }}"
+              echo "agp_actor=${{ github.event.client_payload.agp_actor }}"
+              echo "agp_run_id=${{ github.event.client_payload.agp_run_id }}"
+              echo "agp_artifact_name=${{ github.event.client_payload.agp_artifact_name }}"
+              echo "agp_commit_message=${{ github.event.client_payload.agp_commit_message }}"
+              echo "breaking=${{ github.event.client_payload.breaking }}"
+            } >> "$GITHUB_OUTPUT"
           fi
 
       - uses: actions/checkout@v4
@@ -107,10 +111,9 @@ jobs:
           PY
 
       - name: Regenerate SDK
-        working-directory: python
         run: |
-          uv run --no-project --with pyyaml python scripts/filter_openapi.py \
-            /tmp/agp-schema/openapi.json filter.yaml /tmp/openapi.filtered.json
+          uv run --no-project python scripts/prepare_openapi.py \
+            /tmp/agp-schema/openapi.json /tmp/openapi.filtered.json
           rm -rf src/agent_platform/
           openapi-python-client generate \
             --path /tmp/openapi.filtered.json \
@@ -128,9 +131,8 @@ jobs:
       - name: Verify SDK installs and unit tests pass
         run: |
           python3 -m venv /tmp/sdk-check
-          /tmp/sdk-check/bin/pip install --quiet ./python "pytest>=8,<9"
+          /tmp/sdk-check/bin/pip install --quiet -e . "pytest>=8,<9"
           /tmp/sdk-check/bin/python -c "from agent_platform import Client; Client(api_key='hk-smoke', base_url='http://x'); print('SDK install + import OK')"
-          cd python
           /tmp/sdk-check/bin/pytest tests/ -m "not integration" -q
           /tmp/sdk-check/bin/pytest tests/integration --collect-only -m integration -q
 
@@ -147,6 +149,32 @@ jobs:
             git diff --staged --stat | tail -20 >> "$GITHUB_STEP_SUMMARY"
           fi
 
+      - name: Bump SDK version
+        id: version
+        if: steps.diff.outputs.has_diff == 'true'
+        env:
+          BREAKING: ${{ steps.inputs.outputs.breaking }}
+        run: |
+          # Breaking upstream change -> major bump; any other diff is additive -> minor.
+          NEW_VERSION=$(python3 - <<'PY'
+          import os, re, pathlib
+
+          path = pathlib.Path("pyproject.toml")
+          text = path.read_text()
+          match = re.search(r'^version = "(\d+)\.(\d+)\.(\d+)"', text, re.MULTILINE)
+          major, minor, patch = (int(g) for g in match.groups())
+          if os.environ["BREAKING"] == "true":
+              major, minor, patch = major + 1, 0, 0
+          else:
+              minor, patch = minor + 1, 0
+          new = f"{major}.{minor}.{patch}"
+          path.write_text(text[: match.start()] + f'version = "{new}"' + text[match.end() :])
+          print(new)
+          PY
+          )
+          echo "new_version=$NEW_VERSION" >> "$GITHUB_OUTPUT"
+          echo "Bumped SDK version to $NEW_VERSION (breaking=$BREAKING)" >> "$GITHUB_STEP_SUMMARY"
+
       - name: Open sync PR
         if: steps.diff.outputs.has_diff == 'true'
         uses: peter-evans/create-pull-request@v6
@@ -155,17 +183,18 @@ jobs:
           branch: sync/agent-platform-${{ steps.inputs.outputs.agp_short_sha }}
           delete-branch: true
           commit-message: |
-            chore(sdk): sync to agent_platform@${{ steps.inputs.outputs.agp_short_sha }}
+            chore(sdk): sync to agent_platform@${{ steps.inputs.outputs.agp_short_sha }} (v${{ steps.version.outputs.new_version }})
 
             Upstream commit: ${{ steps.inputs.outputs.agp_commit_message }}
             By: ${{ steps.inputs.outputs.agp_actor }}
-          title: "chore(sdk): sync to agent_platform@${{ steps.inputs.outputs.agp_short_sha }}"
+          title: "chore(sdk): sync to agent_platform@${{ steps.inputs.outputs.agp_short_sha }} (v${{ steps.version.outputs.new_version }})"
           labels: |
             auto-sync
             ${{ steps.inputs.outputs.breaking == 'true' && 'breaking-change' || '' }}
           body: |
             Auto-generated by `regen-on-dispatch.yaml` in response to a backend schema change.
 
+            **Version:** `${{ steps.version.outputs.new_version }}` (${{ steps.inputs.outputs.breaking == 'true' && 'major — breaking upstream change' || 'minor — additive change' }})
             **Upstream:** [agent_platform@${{ steps.inputs.outputs.agp_short_sha }}](https://github.com/hcompai/agent_platform/commit/${{ steps.inputs.outputs.agp_sha }})
             **Triggered by:** ${{ steps.inputs.outputs.agp_actor }}
             **Breaking change:** `${{ steps.inputs.outputs.breaking }}`
@@ -178,6 +207,6 @@ jobs:
 
             ## Reviewer checklist
 
-            - [ ] Diff in `python/src/` looks consistent with upstream changes
-            - [ ] If labelled `breaking-change`: SemVer major bump justified, downstream consumers notified
+            - [ ] Diff in `src/` looks consistent with upstream changes
+            - [ ] Version bump (`${{ steps.version.outputs.new_version }}`) matches the change's severity
             - [ ] Tests pass (CI)
