+++ .github/workflows/test.yaml
@@ -2,16 +2,14 @@ name: Test
 
 on:
   pull_request:
-    paths:
-      - "python/**"
-      - "openapi.json"
-      - ".github/workflows/test.yaml"
+    paths-ignore:
+      - "README.md"
+      - "**/*.md"
   push:
     branches: [main]
-    paths:
-      - "python/**"
-      - "openapi.json"
-      - ".github/workflows/test.yaml"
+    paths-ignore:
+      - "README.md"
+      - "**/*.md"
 
 permissions:
   contents: read
@@ -23,28 +21,42 @@ jobs:
     steps:
       - uses: actions/checkout@v4
 
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
+
+      - name: Install openapi-python-client
+        run: uv tool install openapi-python-client==0.28.3
+
+      - name: Regenerate SDK from committed openapi.json
+        run: |
+          uv run --no-project python scripts/prepare_openapi.py \
+            openapi.json /tmp/openapi.filtered.json
+          rm -rf src/agent_platform/
+          openapi-python-client generate \
+            --path /tmp/openapi.filtered.json \
+            --config config.yaml \
+            --custom-template-path templates/ \
+            --meta uv \
+            --output-path src/ \
+            --overwrite
+          cp templates/__init__.py.static src/agent_platform/__init__.py
+          rm /tmp/openapi.filtered.json
+
       - name: Install SDK + dev deps
-        working-directory: python
         run: |
           python3 -m venv /tmp/sdk-venv
-          /tmp/sdk-venv/bin/pip install --quiet -e ./. "pytest>=8,<9" "python-dotenv>=1,<2"
+          /tmp/sdk-venv/bin/pip install --quiet -e . "pytest>=8,<9" "python-dotenv>=1,<2"
 
       - name: Smoke import
-        run: |
-          /tmp/sdk-venv/bin/python -c "from agent_platform import Client, AsyncClient; Client(api_key='hk-smoke', base_url='http://x'); print('SDK install + import OK')"
+        run: /tmp/sdk-venv/bin/python -c "from agent_platform import Client, AsyncClient; Client(api_key='hk-smoke', base_url='http://x'); print('SDK install + import OK')"
 
-      - name: Unit tests (model round-trip + structural)
-        working-directory: python
+      - name: Unit tests
         run: /tmp/sdk-venv/bin/pytest tests/ -m "not integration" -v
 
-      - name: Collect integration tests (catches stale imports without hitting the API)
-        working-directory: python
+      - name: Collect integration tests
         run: /tmp/sdk-venv/bin/pytest tests/integration --collect-only -m integration -q
 
   integration-live:
-    # Hits the live prod backend with HAI_API_KEY_TEST. Only runs when the secret
-    # is configured (forks / contributors without secrets skip this safely).
-    # Tests are namespaced (sdkit-<ts>-<hex>) so artefacts on failure are easy to spot.
     if: ${{ github.event_name == 'pull_request' && github.event.pull_request.head.repo.full_name == github.repository }}
     runs-on: ubuntu-latest
     timeout-minutes: 20
@@ -55,15 +67,33 @@ jobs:
     steps:
       - uses: actions/checkout@v4
 
-      - name: Install SDK + dev deps
-        working-directory: python
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
+
+      - name: Install openapi-python-client
+        run: uv tool install openapi-python-client==0.28.3
+
+      - name: Regenerate SDK from committed openapi.json
+        run: |
+          uv run --no-project python scripts/prepare_openapi.py \
+            openapi.json /tmp/openapi.filtered.json
+          rm -rf src/agent_platform/
+          openapi-python-client generate \
+            --path /tmp/openapi.filtered.json \
+            --config config.yaml \
+            --custom-template-path templates/ \
+            --meta uv \
+            --output-path src/ \
+            --overwrite
+          cp templates/__init__.py.static src/agent_platform/__init__.py
+          rm /tmp/openapi.filtered.json
+
+      - name: Install SDK
         run: |
           python3 -m venv /tmp/sdk-venv
-          /tmp/sdk-venv/bin/pip install --quiet \
-            --index-url https://pypi.org/simple \
-            -e ./. "pytest>=8,<9" "python-dotenv>=1,<2"
+          /tmp/sdk-venv/bin/pip install --quiet --index-url https://pypi.org/simple \
+            -e . "pytest>=8,<9" "python-dotenv>=1,<2"
 
       - name: Run fast integration tier
         if: env.HAI_API_KEY_TEST != ''
-        working-directory: python
         run: /tmp/sdk-venv/bin/pytest tests/integration -m "integration and not slow" -v --tb=short
