+++ config.yaml
@@ -8,6 +8,10 @@ package_version_override: "0.1.0"
 use_path_prefixes_for_title_model_names: true
 field_prefix: "field_"
 post_hooks:
+  # Strip the standalone-package skeleton that `--meta uv` writes alongside the
+  # generated package. We install + ship via the root pyproject.toml; these
+  # would just be unused noise in every regen diff.
+  - "rm -f pyproject.toml README.md .gitignore"
   # I    = isort (group/sort imports)
   # F811 = remove redundant re-imports that the Jinja templates can still emit
   # Pinning via uvx makes the regen reproducible: bare `ruff` resolves to whatever
