+++ pyproject.toml
@@ -1,11 +1,3 @@
-# Installable from-repo: `pip install ./python` and `uv pip install ./python` (from the
-# agent-api repo root). PyPI publish is wired up via the `publish.yaml` workflow once the
-# repo is made public.
-#
-# Python 3.10+ is required because openapi-python-client 0.28.x itself requires 3.10+ to run.
-# Aligning the SDK's runtime requirement avoids the split where someone could install the SDK
-# but not regenerate it locally.
-
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
@@ -22,12 +14,13 @@ dev = [
 
 [project]
 name = "agent-platform"
-version = "0.1.0"
+version = "0.0.0"
 description = "Python SDK for H Company's Agent Platform"
 requires-python = ">=3.10"
 readme = "README.md"
-authors = [{name = "H Company", email = "engineering@hcompany.ai"}]
-license = {text = "Proprietary"}
+authors = [{name = "H Company"}]
+license = {text = "MIT"}
+license-files = ["LICENSE"]
 dependencies = [
   "httpx>=0.23.0,<0.29.0",
   "pydantic>=2.0,<3",
