--- python/filter.yaml
@@ -1,27 +1,0 @@
-# SDK generation filter
-#
-# Endpoints listed here are removed from openapi.json BEFORE the SDK is
-# generated by openapi-python-client. Apply via scripts/filter_openapi.py:
-#
-#     python scripts/filter_openapi.py openapi.json filter.yaml openapi.filtered.json
-#
-# The schema-sync workflow runs this automatically. To re-enable an endpoint,
-# delete its entry here and rerun the filter (the next regen will include it).
-#
-# Schemas referenced only by excluded paths are removed automatically — no
-# need to list them here.
-
-# Whole paths to drop (every method on these paths is removed).
-#exclude_paths:
-  #- /api/v2/environments
-  #- /api/v2/environments/{env_identifier}
-  #- /api/v2/memories
-  #- /api/v2/memories/{id}
-
-# Specific method+path combinations to drop while keeping other methods on
-# the same path. Methods are case-insensitive.
-exclude_operations:
-  - method: put
-    path: /api/v2/agents/{agent_identifier}
-  - method: delete
-    path: /api/v2/agents/{agent_identifier}
