+++ tests/integration/test_auth_smoke.py
