+++ tests/integration/test_session_browser.py
